add branch 'dev' create package main and test The main package contains the entire kiosk. Currently, we do not organize the package separately
Later, when the abstraction process is over, I will try to separate the package.

The test package was created to test the business logic.

It was decided that it was better to create a test class separately and then modify and add it than to test it in the main menu class.

Business logic implements a shopping cart function,

I tried to operate the shopping basket function lightly in the main menu class.